+++ aniso8601/time.py
@@ -6,7 +6,8 @@
 # This software may be modified and distributed under the terms
 # of the BSD license.  See the LICENSE file for details.
 
-from aniso8601.builder import TupleBuilder, PythonTimeBuilder
+from aniso8601.builders import TupleBuilder
+from aniso8601.builders.pythontimebuilder import PythonTimeBuilder
 from aniso8601.date import parse_date
 from aniso8601.exceptions import ISOFormatError
 from aniso8601.resolution import TimeResolution